Current board wrap functionality first commit
This commit includes all the current board wrap functionality. Currently in this branch the shadow functionality is disabled and when dropping a piece you are able to move the piece through the left or right border on any orientation however you cannot rotate the piece in between borders and you cannot rotate a piece after having moved through a border once.